requirements.txt
# Main dependencies
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
django==4.2.21
django-cors-headers==3.13.0
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
idna==3.4
loguru==0.6.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
win32-setctime==1.1.0

# Dev dependencies
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.1
attrs==22.2.0
black==22.12.0
cfgv==3.3.1
chardet==4.0.0
click==8.1.3
colorama==0.4.6
coverage==7.1.0
distlib==0.3.6
filelock==3.9.0
identify==2.5.17
isort==5.12.0
more-itertools==9.0.0
multidict==6.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==0.13.1
pre-commit==2.21.0
py==1.11.0
tomli==2.0.1
typing-extensions==4.4.0
virtualenv==20.17.1
wcwidth==0.2.6
pytest==5.4.3
